# Handle restricted dependencies as implicit multiple-constraints dependencies