--- a/tests/installation/fixtures/update-with-locked-extras.test
+++ b/tests/installation/fixtures/update-with-locked-extras.test
@@ -8,7 +8,7 @@ files = []
 
 [package.dependencies]
 "B" = {version = "^1.0", optional = true}
-"C" = {version = "^1.0", markers = "python_version >= \"2.7\" and python_version < \"2.8\""}
+"C" = {version = ">=1.0,<2.0", markers = "python_version >= \"2.7\" and python_version < \"2.8\""}
 
 [package.extras]
 foo = ["B"]