@dimbleby 2024-07-23
> I left some notes on potential performance improvements; perhaps it could help speed things up :)

I think it is likely that you are micro-optimizing essentially irrelevant parts of the code.  If you want to make performance improvements - recommend that the first thing to do is to profile, so that you spend your time optimizing the right things

But perhaps I am wrong, and you are now seeing results much better than those in the comment at the top of the thread?  If so - submit a merge request!