@jorenham 2024-07-23
> > I left some notes on potential performance improvements; perhaps it could help speed things up :)
> 
> I think it is likely that you are micro-optimizing essentially irrelevant parts of the code. 

I don't agree that improvements to the runtime complexity are the same as "micro-optimizing". 

Plus, my suggestions will also result in fewer lines of code, without harming readability. So even if the performance benefits are minimal, at the very least there are no disadvantages.